refactor(ICache): move utils into a seperate file
